msm_serial_hs_lite: Handle Tx request during System Suspend
A corner case is occuring, that performing un-clock register
access and leading to Kernel Crash

Corner Case:
When system is in suspend state, Tx operation is getting called
from the client driver by a fflush system call.

To overcome this case, added a conditional check in the driver
while Tx operation. If system is in suspend state, return back
immediately.

CRs-Fixed: 570113
Change-Id: I693693fa9cc3b4b76ed5c7b6c7f4b6f7547e86fa
Signed-off-by: Dilip Kota <c_dkota@codeaurora.org>