Revert "tcp: do not lock listener to process SYN packets"
This reverts commit 69c07f49a8223b53f498ef07183ed45a9369e89b.

This commit belongs to the patch set (https://lwn.net/Articles/659199/)
that attempts to remove the use of locks on the socket table by
relocating the SYN table to a separate hash table and adding a spin lock
to protect the SYN request queue. Adding only this commit introduces a
race condition for our V3.4 kernel for TCP listens, since the TCP SYN
data structures can be corrupted.

Change-Id: Ia7969aeedae411677b307a8e094f9a4cc02b801d